flask/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.7
Flask==2.2.3
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
langchain==0.0.147
linecache2==1.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python

mysqlclient==2.1.1

numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.7
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tqdm==4.65.0
traceback2==1.4.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unittest2==1.1.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.9.1
google-search-results==2.4.2